session: move get_session_path to commands.rs
