+++ alvr/common/src/data/legacy_session.rs
@@ -1,5 +1,5 @@
 use super::{legacy_settings as settings, Settings};
-use crate::prelude::*;
+use crate::{commands, prelude::*};
 use serde::{Deserialize, Serialize};
 use serde_json as json;
 use settings_schema_legacy::SchemaNode;
@@ -14,17 +14,6 @@ use std::{
 // SessionSettings is similar to Settings but it contains every branch, even unused ones. This is
 // the settings representation that the UI uses.
 type SessionSettings = settings::SettingsDefault;
-
-pub fn get_session_path(base: &Path) -> StrResult<PathBuf> {
-    if cfg!(windows) {
-        Ok(base.join("session.json"))
-    } else {
-        Ok(dirs::config_dir()
-            .ok_or_else(|| "no config_dir found")?
-            .join("alvr")
-            .join("session.json"))
-    }
-}
 
 pub fn load_session(path: &Path) -> StrResult<SessionDesc> {
     trace_err!(json::from_str(&trace_err!(fs::read_to_string(path))?))
@@ -571,7 +560,11 @@ impl DerefMut for SessionLock<'_> {
 
 impl Drop for SessionLock<'_> {
     fn drop(&mut self) {
-        save_session(self.session_desc, &get_session_path(&self.dir).unwrap()).unwrap();
+        save_session(
+            self.session_desc,
+            &commands::get_session_path(&self.dir).unwrap(),
+        )
+        .unwrap();
         log_event(Event::SessionUpdated);
     }
 }
@@ -583,7 +576,7 @@ pub struct SessionManager {
 
 impl SessionManager {
     pub fn new(dir: &Path) -> Self {
-        let session_path = get_session_path(&dir).unwrap();
+        let session_path = commands::get_session_path(&dir).unwrap();
         let session_desc = match fs::read_to_string(&session_path) {
             Ok(session_string) => {
                 let json_value = json::from_str::<json::Value>(&session_string).unwrap();
